Conflicts:
	target/m2e-wtp/web-resources/META-INF/maven/com.luuzun/ksca/pom.properties

+++ src/main/java/com/luuzun/ksca/controller/AreaController.java
@@ -10,7 +10,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -19,6 +18,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import com.luuzun.ksca.domain.Area;
 import com.luuzun.ksca.domain.Manager;
 import com.luuzun.ksca.service.AreaService;
+import com.luuzun.ksca.service.SccService;
 
 @Controller
 @RequestMapping("/area/*")
@@ -26,7 +26,8 @@ public class AreaController {
 	private static final Logger logger = LoggerFactory.getLogger(AreaController.class);
 	
 	@Inject	private AreaService service;
-	
+	@Inject	private SccService sccService;
+
 	//Area List
 	@RequestMapping(value="/areaList")
 	public String areaListGET(Model model, HttpSession session, RedirectAttributes rttr) throws Exception{
@@ -101,9 +102,14 @@ public class AreaController {
 		//99 삭제 불가
 		String checkBranch = code.split("-")[2];
 		if(checkBranch.equals("99")){
-			return "ERROR";
+			return "ERROR:def";
 		}
-			
+		
+		//분회에 소속된 경로당이 존재하면 삭제 불가
+		if(sccService.countByArea(code)!=0) {
+			return "ERROR:cascade";
+		}
+		
 		service.delete(code);
 		return code;
 	}
